requirements.txt
Flask==2.3.2
Flask-Session==0.5.0
Werkzeug==2.3.7
python-dotenv==1.0.1
requests==2.31.0
bcrypt==4.1.3
TypeAliases==1.4.0